Add routing and load balancer detail page
This adds the routing to allow navigating between angular pages as
well as the beginning of the load balancer detail page. The detail
page is very basic but is based on both the existing angular Images
detail page and the designs from invision. It includes two empty
tabs that will eventually hold the tables for Pools and Members.

Partially-Implements: blueprint horizon-lbaas-v2-ui
Change-Id: Icbf4396238036b20712ca643e635dd84fe66a86e