Update install instruction for LBaaS v2 dashboard
The suggested way to enable a dashboard is to copy
the 'enabled' file into the dashboard directory.

This commit also enables AngularJS based dashboard default.
The legacy dashboard cannot create a load balancer,
so AngularJS based dashboard is feature compatible now
and more features are coming to AngularJS based implementation.

Change-Id: I4fdc9f56513f2913ca298c52b585291c28f9b54c