Add support for TERMINATED_HTTPS protocol
This adds support for the TERMINATED_HTTPS listener protocol when
creating a new listener. When this option is selected the SSL
Certificates tab is displayed after the Listener Details tab and
allows selecting one or more available certificates. The user must
have barbican available and authority to list certificates and
secrets. Certificate containers must be created in barbican before
they will be available when creating a listener.

Partially-Implements: blueprint horizon-lbaas-v2-ui
Change-Id: Ia9312fa865d85ca977c1daea347d97bd69e9c5ba