Add delete load balancer actions
This adds the single and multiple load balancer delete actions. A
load balancer must be in ACTIVE or ERROR state in order for the
delete action to be available. If a load balancer has a listener the
delete request will currently fail, but future work should allow for
the listener and any related resources to also be deleted.

Partially-Implements: blueprint horizon-lbaas-v2-ui
Change-Id: I3902587089963318c49a969f343e5043d8bdaf50