Improve the technique for generating default resource names
This change improves the way that default names are generated when creating
a new loadbalancer or listener resource.  Currently, it is possible that a
name such as 'Load Balancer 1' is generated as a default name suggestion,
even when there already exists one or more load balancers.  The suggested
default name in these create wizards should be a little smarter and at
least offer a suggested name that makes sense with respect to the number
of loadbalancers that already exist.

Closes-Bug: #1563046
Change-Id: I2a2fe5849b3f8d67efcd693a178f7466af7aef0a