Show the member status properties
This change adds the 'Operating Status' and 'Provisioning Status'
properties to the member attributes that are visible in the
members table and detail pages.  The member statuses are taken
from the Load Balancer status tree API, which returns all child
resources with their corresponding statuses for a particular Load
Balancer.

Closes-Bug: #1563444
Change-Id: Icd812bdc8774c3dd4522ff923229bfecaeb1b9dd