Add loading and error status to detail pages
This adds a loading indicator to detail pages and makes sure that the
page content is not displayed until fully loaded. This prevents the
pages from displaying with partial content and having the updated
values pop in. It also adds an error indicator with a message when
there is an error loading the page. The page content will not display
if there is an error. This prevents the user from seeing the page in
an error state with missing information and broken links.

Change-Id: If26553b0650f38b92e9624baa73e531714baeef3
Closes-Bug: #1561102