Fix auth url for Barbican client
This patch fixes up the auth url if an invalid or no version prefix is
set in OPENSTACK_KEYSTONE_URL. An invalid prefix could be /v2.0 endpoint
when running Keystone v3.

keystoneclient.auth.token_endpoint is deprecated as of the 2.1.0 release
in favor of keystoneauth1.token_endpoint.Token. Reason why it is being
used here is because keystoneauth1 is not in our requirements.txt
(although it is a dependency pulled by required libraries) making this
patch not backportable if we were to add it now. A follow up patch
should handle this deprecation.

Change-Id: I060826fab5067a816297d01d5de20457fd78a742
Story: 2007238
Task: 38576
Task: 29724
(cherry picked from commit fdc34d2087c55bf59c8a20eb2dae13df04157773)